Refactor test_inference
